mishegos: Construct VEX/XOP/EVEX opcodes
The VEX/XOP/EVEX "prefixes" actually behave like being part of the
opcode (the opcode byte must immediately follow the "prefix").
Explicitly constructing these opcodes significantly increases the
likelihood of correctly hitting vector instructions.

This is interesting not only for coverage, but also for the plenty of UD
constraints in AVX-512.